refactor: migrate product page to app router
