+++ src/app/product/[handle]/page.tsx
@@ -1,0 +1,107 @@
+import { headers } from 'next/headers';
+import { notFound } from 'next/navigation';
+import { shopifyFetch } from '@/lib/shopify';
+import ProductClient, { Product } from '@/components/ProductClient';
+
+interface PageProps {
+  params: { handle: string };
+}
+
+export default async function Page({ params }: PageProps) {
+  const header = headers();
+  const isAuthenticated = header.get('x-customer-authenticated') === 'true';
+
+  const query = `
+    query ProductByHandle($handle: String!) {
+      productByHandle(handle: $handle) {
+        id
+        title
+        descriptionHtml
+        priceRange {
+          minVariantPrice {
+            amount
+          }
+        }
+        images(first: 5) {
+          edges {
+            node {
+              url
+              altText
+            }
+          }
+        }
+        variants(first: 10) {
+          edges {
+            node {
+              id
+              title
+              price {
+                amount
+              }
+              availableForSale
+              quantityAvailable
+              selectedOptions {
+                name
+                value
+              }
+            }
+          }
+        }
+        metafields(identifiers: [
+          { namespace: "custom", key: "alloy" },
+          { namespace: "custom", key: "metal" },
+          { namespace: "custom", key: "metal_colour" },
+          { namespace: "custom", key: "thread_type" },
+          { namespace: "custom", key: "gem_type" },
+          { namespace: "custom", key: "gem_colour" },
+          { namespace: "custom", key: "name" },
+          { namespace: "custom", key: "title" },
+          { namespace: "custom", key: "sku" },
+          { namespace: "custom", key: "width" },
+          { namespace: "custom", key: "height" },
+          { namespace: "custom", key: "length" },
+          { namespace: "custom", key: "gauge" },
+          { namespace: "custom", key: "sold_as" },
+          { namespace: "custom", key: "shipping" },
+          { namespace: "custom", key: "base_size" },
+          { namespace: "custom", key: "variants" },
+          { namespace: "custom", key: "variant_label" },
+          { namespace: "custom", key: "fitting" }
+        ]) {
+          key
+          value
+        }
+      }
+    }
+  `;
+
+  const data = await shopifyFetch({ query, variables: { handle: params.handle } });
+
+  if (!data.productByHandle) {
+    notFound();
+  }
+
+  const product: Product = data.productByHandle;
+
+  return <ProductClient product={product} isAuthenticated={isAuthenticated} />;
+}
+
+export async function generateStaticParams() {
+  const query = `
+    {
+      products(first: 250) {
+        edges {
+          node {
+            handle
+          }
+        }
+      }
+    }
+  `;
+
+  const data = await shopifyFetch({ query });
+
+  return data.products.edges.map(({ node }: { node: { handle: string } }) => ({
+    handle: node.handle,
+  }));
+}
